"Add Dockerization and Deployment Configuration for aznopoly-web
This commit includes the addition of necessary files to containerize the aznopoly-web application using Docker and to deploy it on a Kubernetes cluster orchestrated by Helm. A Dockerfile is added to define the app's Docker image, along with a .dockerignore file for excluding unnecessary files from the Docker context. Helm charts are added under helm-charts/ directory to manage the deployment, service, and optionally an ingress and a Horizontal Pod Autoscaler (HPA) for the application on a Kubernetes cluster. It also involves setting up a GitHub Actions workflow for continuously building the Docker image and deploying it to the Kubernetes cluster upon push to the master branch."