Refactor client class and improve error handling
The AzNopolyClient class in client.ts has been refactored for better error handling and clearer code. Previously, the WebSocket socket was initialised within the constructor method, could lead to a null reference error if the socket was accessed before it's initialised. Now, the socket is an optional property of the class and initialised within the connect method improving code readability and reducing complexity. The uuid property in RoomWelcomeEvent interface has been renamed to data for better understanding of its purpose. Additionally, method return types were added for better clarity and understanding of the code.